+++ src/daf_gui/tabs/__init__.py
@@ -1,0 +1,1 @@
+from .schema import *
